+++ lclPy/src/local_search/local_search.rs
@@ -17,11 +17,11 @@ pub trait LocalSearch: Send {
     /// tuple.1 = best score found
     /// tuple.2 = current score
     /// tuple.3 = #iterations
-    fn run(&mut self, log: bool) -> Vec<(u128, isize, isize, usize)>;
+    fn run(&mut self, log: bool) -> Vec<(u128, f64, f64, usize)>;
 
     fn set_problem(&mut self, problem: &Arc<Mutex<dyn Problem>>);
 
     fn set_termination(&mut self, termination: &TerminationFunction);
 
-    fn set_goal(&mut self, minimize:bool);
+    fn set_goal(&mut self, minimize: bool);
 }
